Fix amount decimal truncation and bank_code persistence
- DriverService: cast amount to (float) instead of (int) to preserve decimals
- GenericForm: remove default guard in loadPersistedValues so persist:true
  fields restore from localStorage even when a YAML default exists

+++ src/Services/DriverService.php
@@ -80,7 +80,7 @@ class DriverService
         
         return [
             'code' => $voucher->code,
-            'amount' => (int) ($instructions->cash->amount ?? 0),
+            'amount' => (float) ($instructions->cash->amount ?? 0),
             'currency' => $instructions->cash->currency ?? 'PHP',
             'owner_name' => $voucher->owner->name ?? 'Unknown',
             'base_url' => url(''),
